Refactored all moduels such that they are now working for examples
