--- AdjHE/Covbat.json
@@ -1,20 +1,0 @@
-{
-  "PC" : "/panfs/roc/groups/3/rando149/coffm049/ABCD/Results/01_Gene_QC/filters/filter1/Eigens/EUR_no_rels.eigenvec",   
-  "covar" : "/panfs/roc/groups/3/rando149/coffm049/ABCD/Results/02_Phenotypes/Covars.tsv",
-  "prefix" : "/panfs/roc/groups/3/rando149/coffm049/ABCD/Results/01_Gene_QC/filters/filter1/GRMs/no_rels/no_rels",
-  "pheno" : "/panfs/roc/groups/3/rando149/coffm049/ABCD/Results/02_Phenotypes/asegs.phen",
-  "out" : "/panfs/roc/groups/3/rando149/coffm049/ABCD/Results/03_heritability/Global_volume/Covbat_all.csv",
-  "npc"  : [20],
-  "mpheno" : "all",
-  "k" : 0,
-  "std" :  false,
-  "fixed_effects" : ["age", "female", "household_income"],
-  "argfile" : "/home/christian/Research/Stat_gen/tools/Basu_herit/Example/Arg_file.json",
-  "loop_covars" : false,
-  "covar_relates" : false,
-  "random_groups" : null,
-  "Naive" : false,
-  "Method" : "Covbat",
-  "ids" : null
-}
-
