+++ AdjHE/examples/AdjHE_results
@@ -1,4 +1,3 @@
-h2,ss,var(h2),Analysis time,mem
-0.7771357704709193,0.0,0.0008580242940030975,3.9774824539999827,2633.192
-NA,NA,NA,3.9774824539999827,2633.192
-0.7771357704709193,0.0,0.0008580242940030975,NA,NA
+Pheno,PCs,Covariates,h2,ss,var(h2),time,mem
+Head_size,5,None,0.746104,NA,0.0002930944000000008,32.71029685299982,627.236
+Head_size,10,None,0.7444655,NA,0.00032376604224999985,35.69139191100021,627.236
