+++ AdjHE/examples/AdjHE_results.log
@@ -1,4 +1,5 @@
 root - INFO - These are the list of arguments that were input:
-root - INFO - {'PC': 'Example/Input_files/pcas.eigenvec', 'covar': 'Example/Input_files/covar.txt', 'prefix': 'Example/Input_files/grm', 'pheno': 'Example/Input_files/pheno.phen', 'out': 'Example/AdjHE_results', 'npc': [5, 10], 'mpheno': ['Head_size'], 'k': 0, 'std': False, 'fixed_effects': ['Tidyness', 'Happy_camper'], 'argfile': '/home/christian/Research/Stat_gen/tools/Basu_herit/Example/Arg_file.json', 'loop_covars': False, 'covar_relates': False, 'random_groups': 'abcd_site', 'Naive': True, 'Method': 'GCTA', 'ids': None}
+root - INFO - {'PC': 'AdjHE/examples/Input_files/pcas.eigenvec', 'covar': 'AdjHE/examples/Input_files/covar.txt', 'prefix': 'AdjHE/examples/Input_files/grm', 'pheno': 'AdjHE/examples/Input_files/pheno.phen', 'out': 'AdjHE/examples/AdjHE_results', 'npc': [5, 10], 'mpheno': ['Head_size'], 'k': 0, 'std': False, 'fixed_effects': ['Tidyness', 'Happy_camper'], 'argfile': '/home/christian/Research/Stat_gen/tools/Basu_herit/Example/Arg_file.json', 'loop_covars': False, 'covar_relates': False, 'random_groups': 'abcd_site', 'Naive': True, 'Method': 'GCTA', 'ids': None}
 root - INFO - Loading data
 root - INFO - Estimating
+root - INFO - Writing results to AdjHE/examples/AdjHE_results
