+++ AdjHE/examples/Generic.json
@@ -1,9 +1,9 @@
 {
-  "PC" : "Example/Input_files/pcas.eigenvec",
-  "covar" : "Example/Input_files/covar.txt",
-  "prefix" : "Example/Input_files/grm",
-  "pheno" : "Example/Input_files/pheno.phen",
-  "out" : "Example/AdjHE_results",
+  "PC" : "AdjHE/examples/Input_files/pcas.eigenvec",
+  "covar" : "AdjHE/examples/Input_files/covar.txt",
+  "prefix" : "AdjHE/examples/Input_files/grm",
+  "pheno" : "AdjHE/examples/Input_files/pheno.phen",
+  "out" : "AdjHE/examples/AdjHE_results",
   "npc"  : [5, 10],
   "mpheno" : ["Head_size"],
   "k" : 0,
